Update README.md
The research paper featured in the GeoInfo 2023 Conference Repository, titled "Traffic Congestion Prediction Using Time Series: A Comparison of LSTM and Prophet," delves into the realm of time series analysis for forecasting traffic congestion. This study conducts an in-depth comparison between two prominent methodologies: LSTM (Long Short-Term Memory) and Prophet. For this purpose, it utilizes data sourced from the Performance Measurement System (PeMS) in California.

The findings of this research indicate a superior efficiency of LSTM over Prophet in the context of traffic prediction. Looking ahead, the paper outlines plans for future research, which will explore a variety of other time series prediction models. These models will be specifically applied to the structure of road networks, aiming to enhance the understanding and forecasting of traffic dynamics.